+++ packages/core/src/constants/index.ts
@@ -1,5 +1,6 @@
 import CPU_COLORMAPS from './cpuColormaps';
 import RENDERING_DEFAULTS from './rendering';
 import EPSILON from './epsilon';
+import MPR_CAMERA_VALUES from './mprCameraValues';
 
-export { CPU_COLORMAPS, RENDERING_DEFAULTS, EPSILON };
+export { CPU_COLORMAPS, RENDERING_DEFAULTS, MPR_CAMERA_VALUES, EPSILON };