+++ packages/tools/src/eventListeners/segmentation/segmentationDataModifiedEventListener.ts
@@ -1,6 +1,6 @@
 import { cache } from '@cornerstonejs/core';
 
-import triggerSegmentationRender from '../../utilities/triggerSegmentationRender';
+import triggerSegmentationRender from '../../utilities/segmentation/triggerSegmentationRender';
 import SegmentationRepresentations from '../../enums/SegmentationRepresentations';
 import * as SegmentationState from '../../stateManagement/segmentation/segmentationState';
 import { SegmentationDataModifiedEventType } from '../../types/EventTypes';
@@ -46,7 +46,7 @@ const onSegmentationDataModified = function (
     // Trigger modified on the imageData to update the image
     imageData.modified();
     toolGroupIds =
-      SegmentationState.getToolGroupsWithSegmentation(segmentationId);
+      SegmentationState.getToolGroupIdsWithSegmentation(segmentationId);
   } else {
     throw new Error(
       `onSegmentationDataModified: representationType ${type} not supported yet`