+++ packages/tools/src/stateManagement/segmentation/config/segmentationConfig.ts
@@ -4,6 +4,7 @@ import * as SegmentationState from '../../../stateManagement/segmentation/segmen
 import {
   RepresentationConfig,
   SegmentationRepresentationConfig,
+  SegmentSpecificRepresentationConfig,
 } from '../../../types/SegmentationStateTypes';
 
 /**
@@ -55,7 +56,10 @@ function setGlobalRepresentationConfig(
     ...globalConfig,
     representations: {
       ...globalConfig.representations,
-      [representationType]: config,
+      [representationType]: {
+        ...globalConfig.representations[representationType],
+        ...config,
+      },
     },
   });
 }
@@ -71,6 +75,13 @@ function getToolGroupSpecificConfig(
   return SegmentationState.getToolGroupSpecificConfig(toolGroupId);
 }
 
+/**
+ * Sets the tool group specific configuration for the segmentation
+ * representation. This will apply to all segmentation representations.
+ * @param toolGroupId - The tool group id where the segmentation representation belongs to.
+ * @param segmentationRepresentationConfig - This is the configuration object that you will use to set the default values for
+ * the segmentation representation.
+ */
 function setToolGroupSpecificConfig(
   toolGroupId: string,
   segmentationRepresentationConfig: SegmentationRepresentationConfig
@@ -78,6 +89,91 @@ function setToolGroupSpecificConfig(
   SegmentationState.setToolGroupSpecificConfig(
     toolGroupId,
     segmentationRepresentationConfig
+  );
+}
+
+/**
+ * Give the segmentation representation UID, return the corresponding config
+ * which is shared by all segments in the segmentation representation. This is
+ * an optional level of configuration that can be set by the user, by default
+ * it will fallback to the toolGroup specific config, if not set, it will fallback
+ * to the global config.
+ *
+ * @param segmentationRepresentationUID - The uid of the segmentation representation
+ * @param config - The configuration for the representation. This is an object
+ * only containing the representation type as key and the config as value.
+ * @returns - The configuration for the representation.
+ */
+function getSegmentationRepresentationSpecificConfig(
+  toolGroupId: string,
+  segmentationRepresentationUID: string
+): RepresentationConfig {
+  return SegmentationState.getSegmentationRepresentationSpecificConfig(
+    toolGroupId,
+    segmentationRepresentationUID
+  );
+}
+
+/**
+ * Set the segmentation representation specific configuration for the
+ * segmentation representation. This will apply to all segments in the
+ * segmentation representation and has higher priority than the toolGroup
+ * specific config.
+ *
+ * @param segmentationRepresentationUID - The uid of the segmentation representation
+ * @param config - The configuration for the representation. This is an object
+ * only containing the representation type as key and the config as value.
+ */
+function setSegmentationRepresentationSpecificConfig(
+  toolGroupId: string,
+  segmentationRepresentationUID: string,
+  config: RepresentationConfig
+): void {
+  SegmentationState.setSegmentationRepresentationSpecificConfig(
+    toolGroupId,
+    segmentationRepresentationUID,
+    config
+  );
+}
+
+/**
+ * Get the segment specific configuration for the segmentation representation.
+ *
+ * @param toolGroupId - The tool group id where the segmentation representation belongs to.
+ * @param segmentationRepresentationUID  - The uid of the segmentation representation
+ * @param segmentIndex - The index of the segment
+ * @returns - The configuration for the segment index in the segmentation representation that is shown in the toolGroup's viewport
+ */
+function getSegmentSpecificConfig(
+  toolGroupId: string,
+  segmentationRepresentationUID: string,
+  segmentIndex: number
+): RepresentationConfig {
+  return SegmentationState.getSegmentSpecificRepresentationConfig(
+    toolGroupId,
+    segmentationRepresentationUID,
+    segmentIndex
+  );
+}
+
+/**
+ * Set the segment specific configuration for the segmentation representation.
+ * This configuration, if specified, has higher priority than the segmentation representation specific config,
+ * and the toolGroup specific config. The order of priority is: segment specific config > segmentation representation specific config > toolGroup specific config > global config
+ * @param toolGroupId - The tool group id where the segmentation representation belongs to.
+ * @param segmentationRepresentationUID - The uid of the segmentation representation
+ * @param segmentIndex - The index of the segment
+ * @param config - The configuration for the representation. This is an object
+ */
+function setSegmentSpecificConfig(
+  toolGroupId: string,
+  segmentationRepresentationUID: string,
+  config: SegmentSpecificRepresentationConfig
+): void {
+  SegmentationState.setSegmentSpecificRepresentationConfig(
+    toolGroupId,
+    segmentationRepresentationUID,
+    config
   );
 }
 
@@ -90,4 +186,10 @@ export {
   // ToolGroup Specific
   getToolGroupSpecificConfig,
   setToolGroupSpecificConfig,
+  // segmentation representation specific config
+  getSegmentationRepresentationSpecificConfig,
+  setSegmentationRepresentationSpecificConfig,
+  // segment specific config
+  getSegmentSpecificConfig,
+  setSegmentSpecificConfig,
 };