+++ packages/tools/src/synchronizers/index.ts
@@ -1,9 +1,11 @@
 import createCameraPositionSynchronizer from './synchronizers/createCameraPositionSynchronizer';
 import createVOISynchronizer from './synchronizers/createVOISynchronizer';
 import createZoomPanSynchronizer from './synchronizers/createZoomPanSynchronizer';
+import createStackImageSynchronizer from './synchronizers/createStackImageSynchronizer';
 
 export {
   createCameraPositionSynchronizer,
   createVOISynchronizer,
   createZoomPanSynchronizer,
+  createStackImageSynchronizer,
 };