+++ packages/tools/src/tools/index.ts
@@ -9,6 +9,7 @@ import VolumeRotateMouseWheelTool from './VolumeRotateMouseWheelTool';
 import MIPJumpToClickTool from './MIPJumpToClickTool';
 import CrosshairsTool from './CrosshairsTool';
 import MagnifyTool from './MagnifyTool';
+import ReferenceLinesTool from './ReferenceLinesTool';
 //
 import BidirectionalTool from './annotation/BidirectionalTool';
 import LengthTool from './annotation/LengthTool';
@@ -20,6 +21,7 @@ import PlanarFreehandROITool from './annotation/PlanarFreehandROITool';
 import ArrowAnnotateTool from './annotation/ArrowAnnotateTool';
 import AngleTool from './annotation/AngleTool';
 import ReferenceCursors from './ReferenceCursors';
+import ReferenceLines from './ReferenceLinesTool';
 
 // Segmentation DisplayTool
 import SegmentationDisplayTool from './displayTools/SegmentationDisplayTool';
@@ -48,6 +50,7 @@ export {
   MIPJumpToClickTool,
   // Annotation Tools
   CrosshairsTool,
+  ReferenceLinesTool,
   BidirectionalTool,
   LengthTool,
   ProbeTool,
@@ -67,4 +70,5 @@ export {
   RectangleROIStartEndThresholdTool,
   BrushTool,
   MagnifyTool,
+  ReferenceLines,
 };