+++ packages/tools/src/types/LabelmapTypes.ts
@@ -19,6 +19,10 @@ export type LabelmapConfig = {
   fillAlpha?: number;
   /** alpha of the fill when inactive */
   fillAlphaInactive?: number;
+  /** alpha of outline for active segmentation */
+  outlineOpacity?: number;
+  /** alpha of outline for inactive segmentation */
+  outlineOpacityInactive?: number;
 };
 
 /**